Alarm system with some features v.1
This is the first version of the alarm system with some features.l
The alarm system can be armed with a button and then then ultrasonic sensor is acitivated and waits for any blockage. This activates the alarm that consists of an LED and a buzzer.
Future development: Sends message to thinkspeak (or other) that a break in has occured.
- Seb